frontend/public/image/gamesgroup/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
Django==3.2.11
django-allauth==0.51.0
django-auto-logout==0.5.0
django-clearcache==1.2.1
django-cors-headers==3.13.0
django-crontab==0.7.1
django-decouple==2.1
django-phone-field==1.8.1
django-picklefield==3.0.1
django-redshift-backend==3.0.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-extra-fields==3.4.1
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.0
psycopg2-binary==2.9.5
psycopg2==2.9.4
pandas==1.5.1
django-phone-field==1.8.1
django-picklefield==3.0.1
pycparser==2.21
PyJWT==1.7.1
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.2
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
upgrade-pip==0.1.4
uritemplate==4.1.1
urllib3==1.26.15
